Move PivotOps to separate module
