+++ py-polars/polars/internals/dataframe/frame.py
@@ -46,6 +46,7 @@ from polars.internals.construction import (
     sequence_to_pydf,
     series_to_pydf,
 )
+from polars.internals.dataframe.pivot import PivotOps
 from polars.internals.slice import PolarsSlice
 from polars.utils import (
     _prepare_row_count_args,
@@ -99,11 +100,6 @@ if sys.version_info >= (3, 10):
 else:
     from typing_extensions import TypeAlias
 
-
-# A type variable used to refer to a polars.DataFrame or any subclass of it.
-# Used to annotate DataFrame methods which returns the same type as self.
-DF = TypeVar("DF", bound="DataFrame")
-
 if TYPE_CHECKING:
     from polars.internals.type_aliases import (
         AsofJoinStrategy,
@@ -129,6 +125,10 @@ if TYPE_CHECKING:
     # NOTE: wrapping these as strings is necessary for Python <3.10
     MultiRowSelector: TypeAlias = "slice | range | list[int] | pli.Series"
     MultiColSelector: TypeAlias = "slice | range | list[int] | list[str] | pli.Series"
+
+# A type variable used to refer to a polars.DataFrame or any subclass of it.
+# Used to annotate DataFrame methods which returns the same type as self.
+DF = TypeVar("DF", bound="DataFrame")
 
 
 def wrap_df(df: PyDataFrame) -> DataFrame:
@@ -6750,72 +6750,6 @@ class GroupBy(Generic[DF]):
 
         """
         return self.agg(pli.all().list())
-
-
-class PivotOps(Generic[DF]):
-    """Utility class returned in a pivot operation."""
-
-    def __init__(
-        self,
-        df: PyDataFrame,
-        by: str | list[str],
-        pivot_column: str | list[str],
-        values_column: str | list[str],
-        dataframe_class: type[DF],
-    ):
-        self._df = df
-        self.by = by
-        self.pivot_column = pivot_column
-        self.values_column = values_column
-        self._dataframe_class = dataframe_class
-
-    def first(self) -> DF:
-        """Get the first value per group."""
-        return self._dataframe_class._from_pydf(
-            self._df.pivot(self.by, self.pivot_column, self.values_column, "first")
-        )
-
-    def sum(self) -> DF:
-        """Get the sum per group."""
-        return self._dataframe_class._from_pydf(
-            self._df.pivot(self.by, self.pivot_column, self.values_column, "sum")
-        )
-
-    def min(self) -> DF:
-        """Get the minimal value per group."""
-        return self._dataframe_class._from_pydf(
-            self._df.pivot(self.by, self.pivot_column, self.values_column, "min")
-        )
-
-    def max(self) -> DF:
-        """Get the maximal value per group."""
-        return self._dataframe_class._from_pydf(
-            self._df.pivot(self.by, self.pivot_column, self.values_column, "max")
-        )
-
-    def mean(self) -> DF:
-        """Get the mean value per group."""
-        return self._dataframe_class._from_pydf(
-            self._df.pivot(self.by, self.pivot_column, self.values_column, "mean")
-        )
-
-    def count(self) -> DF:
-        """Count the values per group."""
-        return self._dataframe_class._from_pydf(
-            self._df.pivot(self.by, self.pivot_column, self.values_column, "count")
-        )
-
-    def median(self) -> DF:
-        """Get the median value per group."""
-        return self._dataframe_class._from_pydf(
-            self._df.pivot(self.by, self.pivot_column, self.values_column, "median")
-        )
-
-    def last(self) -> DF:
-        """Get the last value per group."""
-        return self._dataframe_class._from_pydf(
-            self._df.pivot(self.by, self.pivot_column, self.values_column, "last")
-        )
 
 
 class GBSelection(Generic[DF]):
